requirements.txt
Django==6.0.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.2.1
psycopg2-binary==2.9.11
stripe==14.1.0
gunicorn==23.0.0
